# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Software Engineering Course Project",
        "title": "SE Project",
        "contact": {
            "name": "Mahdieh Moghiseh",
            "email": "mahdiehmoghiseh81@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/trainer/price": {
            "put": {
                "description": "Trainer sets the price for a training request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "Set price for a request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Trainer Set Price Data",
                        "name": "TrainerSetPrice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TrainerSetPrice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProgramRequestSetPrice"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/trainer/profile/{id}": {
            "get": {
                "description": "Retrieves the profile information of a trainer by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "Get trainer profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainer profile information",
                        "schema": {
                            "$ref": "#/definitions/dto.TrainerResponse"
                        }
                    },
                    "404": {
                        "description": "Trainer not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the profile information of a trainer by UserID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "Edit trainer profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Trainer profile data",
                        "name": "trainer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TrainerEdit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated trainer profile",
                        "schema": {
                            "$ref": "#/definitions/dto.TrainerResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Trainer not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/trainer/program": {
            "post": {
                "description": "create a training program by trainer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "creates a program",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Trainer Create Program data",
                        "name": "TrainingProgram",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TrainingProgram"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/trainer/program/sport-activity": {
            "put": {
                "description": "add sport activity to program by trainer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "add sport activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add Sport Activity data",
                        "name": "SportActivity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddSportActivity"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/trainer/requests/": {
            "get": {
                "description": "get requests of a trainer by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "Get requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainer requests",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RequestsInTrainerPage"
                            }
                        }
                    },
                    "404": {
                        "description": "Trainer not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/trainer/trainees/": {
            "get": {
                "description": "get trainees of a trainer by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "Get trainees",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainer trainees",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.TraineeInTrainerPage"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request: User ID header missing or invalid token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Trainer not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddSportActivity": {
            "type": "object",
            "properties": {
                "program_id": {
                    "type": "integer"
                },
                "sport_activity": {
                    "$ref": "#/definitions/dto.SportActivitRequest"
                }
            }
        },
...
        "dto.ProgramRequestSetPrice": {
            "type": "object",
            "required": [
                "traineeID",
                "trainerID"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "traineeID": {
                    "type": "integer"
                },
                "trainerID": {
                    "type": "integer"
                }
            }
        },
        "dto.RequestsInTrainerPage": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "trainee_name": {
                    "type": "string"
                }
            }
        },
        "dto.Response": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.SportActivitRequest": {
            "type": "object",
            "properties": {
                "expected_value": {
                    "type": "integer"
                },
                "order_number": {
                    "type": "integer"
                },
                "sport": {
                    "$ref": "#/definitions/dto.SportRequest"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "dto.SportRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "video_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.TraineeInTrainerPage": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.TrainerEdit": {
            "type": "object",
            "required": [
                "achievements",
                "active_days",
                "coach_experience",
                "contact",
                "country",
                "education",
                "language",
                "sport",
                "status",
                "user_name"
            ],
            "properties": {
                "achievements": {
                    "type": "string"
                },
                "active_days": {
                    "type": "array",
                    "items": {
                        "type": "boolean"
                    }
                },
                "coach_experience": {
                    "type": "integer"
                },
                "contact": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "education": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "sport": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/dto.UserEdit"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "dto.TrainerProfileCard": {
            "type": "object",
            "properties": {
                "coach_experience": {
                    "type": "integer"
                },
                "contact": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "dto.TrainerResponse": {
            "type": "object",
            "properties": {
                "achievements": {
                    "type": "string"
                },
                "education": {
                    "type": "string"
                },
                "sports": {
                    "type": "string"
                },
                "trainer_profile_card": {
                    "$ref": "#/definitions/dto.TrainerProfileCard"
                }
            }
        },
        "dto.TrainerSetPrice": {
            "type": "object",
            "required": [
                "price",
                "requestId"
            ],
            "properties": {
                "price": {
                    "type": "integer"
                },
                "rejected": {
                    "type": "boolean"
                },
                "requestId": {
                    "type": "integer"
                },
                "trainerID": {
                    "type": "integer"
                }
            }
        },
        "dto.TrainingProgram": {
            "type": "object",
            "required": [
                "end_date",
                "request_id",
                "start_date"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "request_id": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "trainer_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UserEdit": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "block": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "info_id": {
                    "type": "string"
                },
                "info_type": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "wallet": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}